requirements.txt
colorlog==6.8.2
huggingface-hub==0.23.*
numpy==1.26.*
pandas
openpyxl==3.1.2
PyYAML==6.0.1
torch==2.*
transformers==4.40.*
protobuf
sacremoses
sentencepiece==0.2.0
bitsandbytes>=0.42.0
accelerate==0.28.*